social-archive/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary,pool]
python-jose[cryptography]
passlib[bcrypt]
